+++ src/main/java/com/example/shop/controller/command/impl/DeleteUserFromBlacklistCommand.java
@@ -6,6 +6,7 @@ import com.example.shop.model.service.ServiceException;
 import com.example.shop.model.service.UserService;
 import com.example.shop.model.service.impl.UserServiceImpl;
 import com.example.shop.resource.ConfigurationManager;
+import com.example.shop.resource.MessageManager;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -23,19 +24,24 @@ public class DeleteUserFromBlacklistCommand implements ActionCommand {
         String page;
         HttpSession session = request.getSession();
         try {
-            String stringUserId = request.getParameter(PARAM_NAME_USER_ID);
-            int userId = Integer.parseInt(stringUserId);
-            userService.deleteUserFromBlacklist(userId);
-            List<User> users = (List<User>) session.getAttribute("users");
-            for(int i = 0; i < users.size(); i++) {
-                User user = users.get(i);
-                if(user.getUserId() == userId) {
-                    user.setIsBanned(false);
-                    users.set(i, user);
-                    break;
+            int userId = Integer.valueOf(request.getParameter(PARAM_NAME_USER_ID));
+            String nickname = (String) session.getAttribute("nickname");
+            User activeUser = userService.findUserByNickname(nickname);
+            if(activeUser.getUserId() != userId) {
+                userService.deleteUserFromBlacklist(userId);
+                List<User> users = (List<User>) session.getAttribute("users");
+                for (int i = 0; i < users.size(); i++) {
+                    User user = users.get(i);
+                    if (user.getUserId() == userId) {
+                        user.setIsBanned(false);
+                        users.set(i, user);
+                        break;
+                    }
                 }
+                session.setAttribute("users", users);
+            } else {
+                request.setAttribute("actOnYourselfMessage", MessageManager.getProperty("message.actonyourself"));
             }
-            session.setAttribute("users", users);
             page = (String) session.getAttribute("currentPage");
         } catch (ServiceException e) {
             logger.info("Problems in 'DeleteUserFromBlacklistCommand', redirected to error page");
